Change size of lastTime in examples

+++ examples/Supla_Alarm_Communication/Supla_Alarm_Communication.ino
@@ -58,7 +58,7 @@ void loop()
 
   }
 
-  static uint64_t lastTime = 0;
+  static uint32_t lastTime = 0;
   if(millis() - lastTime > 1000) 
   {
     lastTime = millis();
