+++ examples/Supla_Clock/Supla_Clock.ino
@@ -65,7 +65,7 @@ void loop()
 
   Slave.Update();
   
-  static uint64_t lastTime = 0;
+  static uint32_t lastTime = 0;
   if(millis() - lastTime > 1000) 
   {
     lastTime = millis();
